possible incorrect delinting change
(cherry picked from commit bb05f60021b6647ff15d907eee09348bc3c48e70)

+++ ezgz.hpp
@@ -807,7 +807,7 @@ class DeflateReader {
 					if (code.code < 144) {
 						parent->output.addByte(static_cast<char>(code.code));
 					} else {
-						uint8_t full = ((code.code - 144)) << 1 + 144 + input.getBits(1);
+						uint8_t full = static_cast<uint8_t>(((code.code - 144) << 1) + 144 + input.getBits(1));
 						parent->output.addByte(full);
 					}
 				}
